Extract global update settings
